feat: add support for window scrolling
This adds the ability to use the Window to control the scrolling for a
virtual area rather than that areas scroll. To make this possible two
new options need to be added to `useVirtual`, `onScrollElement` and
`scrollOffsetFn`. These allow the `useVirtualWindow` hook to work
correctly with the existing logic to use Window to scroll the
`parentRef` area.