fix: Fix `initialOffset` not working
The range is initially calculated correctly, but the scroll position of the scrollingElem is 0 (which causes the range to be re-calculated to match the visible area).

The same applies for SSR usage: the scroll position cannot be transmitted from the server, so it needs to be restored on the client.

Fixes https://github.com/TanStack/virtual/issues/346