fix(virtual-core): silently fail instead of throwing an error
Before the this code runs, we already verify that the offset for a
specific index exists. If this is not the case, then we bail by just
using a `return`

However, the code changed in this commit then assumes that the same
conditions hold true. However, this is also running in a `setTimeout`
which means that values could have changed in the meantime (which can
happen due to React re-renders).

In our particular issue in Headless UI, it's because somebody is
filtering a Combobox while typing. This in turn changes the `enabled`
and `count`
from the `useVirtualizer` options.

How we use it in Headless UI:
```
let virtualizer = useVirtualizer({
  enabled: options.length !== 0,
  count: options.length,
  // …
})
```

Another issue is that because of the `setTimeout`, you can't even catch
this error in user land.

So to fix it, once we check if the offset is still available inside of
the `setTimeout`, we will also just silently bail if it results in
`undefined`.